requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.
# Order matters to the pip dependency resolver, so sorting this file
# changes how packages are installed.  New dependencies should be
# added in alphabetical order, however, some dependencies may need to
# be installed in a specific order.
#
# PBR should always appear first
pbr>=1.6
# Horizon Core Requirements
Babel>=1.3
Django<1.9,>=1.7
Pint>=0.5 # BSD
django-babel>=0.3.4
django-compressor>=1.4
django-openstack-auth>=2.0.0
django-pyscss>=2.0.2 # BSD License (2 clause)
eventlet>=0.17.4
httplib2>=0.7.5
iso8601>=0.1.9
kombu>=3.0.7
netaddr!=0.7.16,>=0.7.12
oslo.concurrency>=2.3.0 # Apache-2.0
oslo.config>=2.3.0 # Apache-2.0
oslo.i18n>=1.5.0 # Apache-2.0
oslo.policy>=0.5.0 # Apache-2.0
oslo.serialization>=1.4.0 # Apache-2.0
oslo.utils!=2.6.0,>=2.0.0 # Apache-2.0
pyScss>=1.3.4 # MIT License
python-ceilometerclient>=1.5.0
python-cinderclient>=1.3.1
python-glanceclient>=0.18.0
python-heatclient>=0.3.0
python-keystoneclient!=1.8.0,>=1.6.0
python-neutronclient>=2.6.0
python-novaclient!=2.33.0,>=2.28.1
python-saharaclient>=0.10.0
python-swiftclient>=2.2.0
python-troveclient>=1.2.0
pytz>=2013.6
PyYAML>=3.1.0
six>=1.9.0
XStatic>=1.0.0 # MIT License
XStatic-Angular>=1.3.7 # MIT License
XStatic-Angular-Bootstrap>=0.11.0.2 # MIT License
XStatic-Angular-Gettext>=2.1.0.2 # MIT License
XStatic-Angular-lrdragndrop>=1.0.2.2 # MIT License
XStatic-Bootstrap-Datepicker>=1.3.1.0 # Apache 2.0 License
XStatic-Bootstrap-SCSS>=3 # Apache 2.0 License
XStatic-bootswatch>=3.3.5.3 # MIT License
XStatic-D3>=3.1.6.2 # BSD License (3 clause)
XStatic-Hogan>=2.0.0.2 # Apache 2.0 License
XStatic-Font-Awesome>=4.3.0 # SIL OFL 1.1 License, MIT License
XStatic-Jasmine>=2.1.2.0 # MIT License
XStatic-jQuery>=1.7.2 # MIT License
XStatic-JQuery-Migrate>=1.2.1.1 # MIT License
XStatic-JQuery.quicksearch>=2.0.3.1 # MIT License
XStatic-JQuery.TableSorter>=2.14.5.1 # MIT License
XStatic-jquery-ui>=1.10.1 # MIT License
XStatic-JSEncrypt>=2.0.0.2 # MIT License
XStatic-Magic-Search>=0.2.5.1 # Apache 2.0 License
XStatic-mdi==1.1.70.1 # SIL OPEN FONT LICENSE Version 1.1
XStatic-Rickshaw>=1.5.0 # BSD License (prior)
XStatic-roboto-fontface>=0.4.3.2 # Apache 2.0 License
XStatic-smart-table>=1.4.5.3 # MIT License
XStatic-Spin>=1.2.5.2 # MIT License
XStatic-term.js>=0.0.4 # MIT License

# CONTRIB requirements
influxdb==2.8.0
MySQL-python;python_version=='2.7'
oslo.db>=2.4.1  # Apache-2.0
pyvmomi>=5.5.0.2014.1.1
pymongo
